Added check if the user already liked the post

+++ api/post.py
@@ -101,10 +101,18 @@ def like_post() -> Response:
 
 @app.route('/dislike', methods=['POST'])
 def dislike():
+    """Dislikes a post. 
+    If the user has not liked the post, then the counter is not decremented.
+
+    Returns:
+        Response: whether the post was disliked or not
+    """    
     comm_subs = request.get_json() # get the post data
     post_id: int = comm_subs.get('post_id') # get the post id
     user_id: int = comm_subs.get('user_id') # get the user id
-    likes = LikePost.query.filter_by(post_id=post_id, user_id=user_id).delete() # delete the LikePost instance from the database
-    db.session.commit() # commit the changes to the database
-    likes = LikePost.query.filter_by(post_id=post_id).count() # get the number of likes for the post
-    return make_response(jsonify({'msg': 'You dislike this post', 'total_likes': likes}), 200) # return a response to the user
+    if (LikePost.query.filter_by(post_id=post_id, user_id=user_id).first()): # if the user has liked the post
+        LikePost.query.filter_by(post_id=post_id, user_id=user_id).delete() # delete the LikePost instance from the database
+        db.session.commit() # commit the changes to the database
+        likes = LikePost.query.filter_by(post_id=post_id).count() # get the number of likes for the post
+        return make_response(jsonify({'msg': 'You disliked this post', 'total_likes': likes}), 200) # return a response to the user
+    return make_response(jsonify({'msg': 'You have not liked this post'}), 400) # return a response to the user
